[TASK] @process.joinAndTrim removed
It seems like this 'joinAndTrim' processor is not needed anymore.
It was initially done due to NULL/empty values being included in the
output, which resulted in not the best output (plenty of spaces
between class names) or even completely empty attributes (which in HTML
can have side effects).

Relates: #28