[TASK] Column renamed to GridColumn, ColumnEnd removed + nodes migration
M12.Foundation:Column, M12.Foundation:ColumnEnd naming was not the best.
It's a grid column and we should follow here naming Grid[Something],
like we have for M12.Foundation:GridRow.

In addition, the GridColumnEnd seems to be overkill and not worth
effort of maintaining it. In addition it adds a confusion when user
decides to add more columns: what the heck is that column end?
With this change 'classEndColumn' property is available on each column
and positioned in the inspector at the very end, so it doesn't bother
users too much.